arch/x86_64: Migrate APIC definitions to `arch.h`
Since the APIC definitions are arch-defined, move them to arch.h.

Previously the only users of these macros were some static inlines
exported by an APIC header as part of the interrupt controller XPIC
driver under the form of static inlines. This header's includion
was unnecessarily conditioned under the XPIC driver bieng enabled.
Get rid of this unnecessary header (and this condition) and move the
used routines in their respective C files.
